[fasm] Ints are 32 bits instead of 64 bits, simplified size names

+++ myteo/src/parser/ast.ml
@@ -127,7 +127,7 @@ and Expression : sig
     type t = {
       loc: Loc.t;
       raw: string;
-      value: Int64.t;
+      value: Int32.t;
     }
   end
